Implement shortcut key access

+++ src/state/folderItemEditor/listeners.js
@@ -3,10 +3,10 @@ import { bookmarkRepository, folderItemPreferenceRepository } from '../../reposi
 import * as actionTypes from './actionTypes';
 import * as folderItemPreferencesActionTypes from '../folderItemPreferences/actionTypes';
 
-export default {
-  [actionTypes.SAVE_FOLDER_ITEM_EDITOR]: (action, dispatch, store) => {
-    const { folderItem, folderItemPreference } = action;
+import Bookmark from '../../models/Bookmark';
 
+export default {
+  [actionTypes.SAVE_FOLDER_ITEM_EDITOR]: ({folderItem, folderItemPreference}, dispatch, store) => {
     const folderItemPreferences = store.getState().folderItemPreferences.set(folderItemPreference);
     folderItemPreferenceRepository.save(folderItemPreferences);
     dispatch({
@@ -14,17 +14,30 @@ export default {
       folderItemPreferences
     });
 
-    bookmarkRepository.update(folderItem, () => dispatch({
-      type: actionTypes.SAVED_FOLDER_ITEM_EDITOR,
-      folderItem
-    }));
+    if (folderItem instanceof Bookmark) {
+      bookmarkRepository.update(folderItem, () => dispatch({
+        type: actionTypes.SAVED_FOLDER_ITEM_EDITOR,
+        folderItem
+      }));
+    } else {
+      dispatch({
+        type: actionTypes.SAVED_FOLDER_ITEM_EDITOR,
+        folderItem
+      });
+    }
   },
 
-  [actionTypes.REMOVE_FOLDER_ITEM_EDITOR]: (action, dispatch) => {
-    const { folderItem } = action;
-    bookmarkRepository.remove(folderItem, () => dispatch({
-      type: actionTypes.REMOVED_FOLDER_ITEM_EDITOR,
-      folderItem
-    }))
+  [actionTypes.REMOVE_FOLDER_ITEM_EDITOR]: ({folderItem}, dispatch) => {
+    if (folderItem instanceof Bookmark) {
+      bookmarkRepository.remove(folderItem, () => dispatch({
+        type: actionTypes.REMOVED_FOLDER_ITEM_EDITOR,
+        folderItem
+      }));
+    } else {
+      dispatch({
+        type: actionTypes.REMOVED_FOLDER_ITEM_EDITOR,
+        folderItem
+      });
+    }
   },
 };
